rename verifier test files
